+++ src/business/ResourceItem/RecordBookItem.js
@@ -2,6 +2,7 @@ import React, { Fragment } from 'react'
 import { Observer } from 'mobx-react-lite'
 import { useRouterContext } from 'contexts'
 import timespan from 'utils/timespan'
+import { Button } from 'antd-mobile'
 // import config from 'config'
 
 export default function ({ item }) {
@@ -16,14 +17,19 @@ export default function ({ item }) {
           <div className="full-width-auto full-height">
             <div className="dd-common-alignside">
               <div style={{ fontSize: '1.2rem' }}>{item.title}</div>
+              <Button type="primary" size="small" onClick={(e) => {
+                e.stopPropagation();
+                router.pushView('BookChapter', { bid: item.id, id: item.last_seen_id })
+              }}>继续阅读</Button>
             </div>
             <div style={{ padding: '4px 0', color: 'rgb(146, 145, 145)' }}>{item.uname} · {item.type}</div>
-            <div style={{ color: 'rgb(146, 145, 145)' }}>{timespan(new Date(item.last.createdAt || Date.now()))} · {item.last.title || '最新章节'}</div>
+            <div onClick={(e) => {
+              e.stopPropagation();
+              router.pushView('BookChapter', { bid: item.id, id: item.last_seen_id })
+            }} style={{ color: 'rgb(146, 145, 145)' }}>已读到 {item.last_seen_title} · {timespan(new Date(item.last_seen_ts))}</div>
           </div>
         </div>
       </Fragment>
-    }
-    }
+    }}
   </Observer>
-
 }